+++ src/pages/Shoes/ShoesList/ShoesList.js
@@ -1,11 +1,16 @@
 import { Component } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, withRouter } from 'react-router-dom';
 import { GoPlus } from 'react-icons/go';
 import { FaMinus } from 'react-icons/fa';
+import { FaAngleDoubleLeft } from 'react-icons/fa';
+import { FaAngleDoubleRight } from 'react-icons/fa';
 import ShoesCard from '../ShoesCard/ShoesCard';
 import ShoesFilterButton from '../ShoesList/ShoesFilterButton';
-import ShoesPaginate from '../ShoesPaginate/ShoesPaginate';
+import ShoesPaginateNum from '../ShoesPaginate/ShoesPaginateNum';
+import '../ShoesPaginate/ShoesPaginate.scss';
 import './ShoesList.scss';
+
+const SHOW = 8;
 
 class ShoesList extends Component {
   constructor() {
@@ -13,6 +18,8 @@ class ShoesList extends Component {
     this.state = {
       shoesData: [],
       open: false,
+      PageNo: 1,
+      PageNoMax: 1,
     };
   }
 
@@ -27,16 +34,63 @@ class ShoesList extends Component {
   // }
 
   componentDidMount() {
-    fetch('http://10.58.1.224:8000/products', {
-      method: 'GET',
-    })
-      .then(res => res.json())
-      .then(data => {
-        this.setState({
-          shoesData: data.product_list,
-        });
-      });
+    this.getProductData();
   }
+
+  componentDidUpdate(prevProps, prevState) {
+    if (prevProps.location.search !== this.props.location.search) {
+      fetch(`http://10.58.1.224:8000/products${this.props.location.search}`)
+        .then(res => res.json())
+        .then(data => {
+          this.setState({
+            PageNoMax: data.max_page,
+            shoesData: data.product_list,
+          });
+        });
+    }
+  }
+
+  getProductData = () => {
+    const { PageNo } = this.state;
+    fetch(`http://10.58.1.224:8000/products?PageNo=${PageNo}&Show=${SHOW}`)
+      .then(res => res.json())
+      .then(data =>
+        this.setState({
+          PageNoMax: data.max_page,
+          shoesData: data.product_list,
+        })
+      );
+  };
+
+  clickLeft = () => {
+    const { PageNo } = this.state;
+    let nextPageNum = PageNo;
+
+    if (PageNo > 1) {
+      nextPageNum--;
+      this.setState({ PageNo: nextPageNum });
+    }
+    this.props.history.push(
+      `/shoes?PageNo=${Number(nextPageNum)}&Show=${SHOW}`
+    );
+  };
+
+  clickRight = () => {
+    const { PageNo } = this.state;
+    let nextPageNum = PageNo + 1;
+
+    if (nextPageNum > 1) {
+      this.setState({ PageNo: nextPageNum });
+    }
+    this.props.history.push(
+      `/shoes?PageNo=${Number(nextPageNum)}&Show=${SHOW}`
+    );
+  };
+
+  fetchShoes = number => {
+    this.setState({ PageNo: number });
+    this.props.history.push(`/shoes?PageNo=${number}&Show=${SHOW}`);
+  };
 
   handleClickOpen = () => {
     this.setState({
@@ -44,13 +98,14 @@ class ShoesList extends Component {
     });
   };
 
-  // updateDate = () => {
-  //   const {shoesData} = this.props;
-  //   shoesData.filter(element => ‘a’ === element).length;
-  // }
+
 
   render() {
     const { shoesData } = this.state;
+
+    const PageNoNum = [...Array(Number(this.state.PageNoMax)).keys()];
+
+    const { clickLeft, clickRight } = this;
 
     return (
       <div className="ShoesList">
@@ -78,9 +133,15 @@ class ShoesList extends Component {
 
             <div className="lastFilter">
               <select name="" id="" className="filterSelect">
-                <option value="20개씩 보기">10개씩 보기</option>
-                <option value="40개씩 보기">20개씩 보기</option>
-                <option value="80개씩 보기">40개씩 보기</option>
+                <option value="20개씩 보기" >
+                  10개씩 보기
+                </option>
+                <option value="40개씩 보기">
+                  20개씩 보기
+                </option>
+                <option value="80개씩 보기">
+                  40개씩 보기
+                </option>
               </select>
               <button className="filterButton" onClick={this.handleClickOpen}>
                 FILTE
@@ -114,11 +175,27 @@ class ShoesList extends Component {
             </ul>
           </section>
 
-          <ShoesPaginate />
+          <div className="ShoesPaginate">
+            <ul className="paginateUl">
+              <li className="prevArrow">
+                <FaAngleDoubleLeft onClick={clickLeft} />
+              </li>
+              {PageNoNum.map((num, index) => (
+                <ShoesPaginateNum
+                  key={index}
+                  num={num + 1}
+                  fetchShoes={this.fetchShoes}
+                />
+              ))}
+              <li className="nextArrow">
+                <FaAngleDoubleRight onClick={clickRight} />
+              </li>
+            </ul>
+          </div>
         </div>
       </div>
     );
   }
 }
 
-export default ShoesList;
+export default withRouter(ShoesList);
